requirements.txt
glob2~=0.7
pandas~=1.1.5
spacy~=2.3.4
python-dateutil~=2.7.3